+++ chat-widget/src/components/livechatwidget/common/endChat.ts
@@ -75,6 +75,7 @@ const endChat = async (props: ILiveChatWidgetProps, chatSDK: any, setAdapter: an
     }
 
     // Need to clear these states immediately when chat ended from OC.
+    dispatch({ type: LiveChatWidgetActionType.SET_CUSTOM_CONTEXT, payload: undefined });
     dispatch({ type: LiveChatWidgetActionType.SET_CHAT_TOKEN, payload: undefined });
     dispatch({ type: LiveChatWidgetActionType.SET_LIVE_CHAT_CONTEXT, payload: undefined });